+++ lib/src/datasources/http/datasource_http_delete.dart
@@ -1,4 +1,7 @@
+import 'dart:async';
+
 import 'package:data_shaft/data_shaft.dart';
+import 'package:ncc/ncc.dart';
 import 'package:ncc_data_shaft/src/datasources/datasource_ncc.dart';
 import 'package:ncc_data_shaft/src/driver/ncc_connection_driver.dart';
 
@@ -13,11 +16,24 @@ import 'package:ncc_data_shaft/src/driver/ncc_connection_driver.dart';
 /// The [RemoteObject] defines the resource being deleted, and the
 /// [HttpDataShaftDriver] provides the standard HTTP transport layer.
 /// {@endtemplate}
-abstract base class DatasourceDeleteHttp<
+abstract base class DatasourceHttpDelete<
         RemoteObject extends Codable<Object, RemoteObject>>
     extends DatasourceNccDelete<RemoteObject, HttpDataShaftDriver> {
   /// {@macro datasource_delete_http}
   ///
   /// Requires an [HttpDataShaftDriver] instance to perform the network operation.
-  DatasourceDeleteHttp({required super.driver});
+  DatasourceHttpDelete({required super.driver});
+
+  @override
+  FutureOr<RemoteObject> transformation({
+    required RequestResponse<Response> remoteResponse,
+  });
+
+  @override
+  FutureOr<RemoteObject> checkInformation({
+    required RequestResponse<Response> requestResponse,
+    required Map<String, String>? requestHeaders,
+    required Uri? requestUri,
+    Object? requestBody,
+  });
 }